requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
